backend/requirements.txt
flask
flask-cors
transformers
torch
pycryptodome
gunicorn